Restructuring changes.
