+++ app/components/table/index.ts
@@ -1,1 +1,1 @@
-export * from './table';
+export * from './table.component';
