+++ app/components/table/row/row-modified-fields.factory.ts
@@ -1,7 +1,7 @@
 import {Injectable} from '@angular/core';
 import {Subject} from 'rxjs';
 import * as _ from 'lodash';
-import {TableComponent} from '../table';
+import {TableComponent} from '../table.component';
 
 @Injectable()
 export class RowModifiedFieldsFactory {
